% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{create_repo}
\alias{create_repo}
\title{Craete a GitHub repository for a Bioconductor package}
\usage{
create_repo(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\description{
Craete a GitHub repository for a Bioconductor package
}
\details{
If the repository already exists, then it does not error.
This is detected by hte GitHub API returning a 422 HTTP status code.
For other abnormal status codes it errors.
}
