% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{get_bioc_refs}
\alias{get_bioc_refs}
\alias{async_get_bioc_refs}
\title{Get all git refs of a package from Bioconductor git}
\usage{
get_bioc_refs(pkg)

async_get_bioc_refs(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
A data frame with columns:
\itemize{
\item \code{ref}: git ref name, the first one is \code{HEAD}, if the repository is
not empty.
\item \code{hash}: SHA-1 hash.
}
}
\description{
Get all git refs of a package from Bioconductor git
}
\details{
It errors if the repository does not exist or on any git error.

It uses the dummy git protcol.

\code{async_get_bioc_refs()} is the aynchronous version of \code{get_bioc_refs()}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_bioc_refs("limma")
\dontshow{\}) # examplesIf}
}
\seealso{
Other refs: 
\code{\link{get_all_bioc_refs}()},
\code{\link{get_all_github_refs}()},
\code{\link{get_github_refs}()},
\code{\link{get_missing_refs}()},
\code{\link{read_github_state}()},
\code{\link{write_github_state}()}
}
\concept{refs}
