% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{get_github_refs}
\alias{get_github_refs}
\alias{async_get_github_refs}
\title{Get all git refs from the GitHub Bioconductor mirror}
\usage{
get_github_refs(pkg)

async_get_github_refs(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
A data frame with columns:
\itemize{
\item \code{ref}: git ref name, the first one is \code{HEAD}, if the repository exists
and it is not empty.
\item \code{hash}: SHA-1 hash.
}
}
\description{
Get all git refs from the GitHub Bioconductor mirror
}
\details{
It the repository does not exist, it'll return an empty data frame.

\code{async_get_github_refs()} is the aynchronous version of
\code{get_github_refs()}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_github_refs("limma")
\dontshow{\}) # examplesIf}
}
\seealso{
Other refs: 
\code{\link{get_all_bioc_refs}()},
\code{\link{get_all_github_refs}()},
\code{\link{get_bioc_refs}()},
\code{\link{get_missing_refs}()}
}
\concept{refs}
