% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{get_missing_refs}
\alias{get_missing_refs}
\alias{async_get_missing_refs}
\title{Get git refs that are missing from or are different in the GitHub mirror}
\usage{
get_missing_refs(pkg)

async_get_missing_refs(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
A list with entries:
\itemize{
\item \code{missing}: character vector of refs missing from GitHub.
\item \code{extra}: character vector of extra refs on GitHub.
\item \code{different}: character vecotr of refs that have a different hash at
GitHub.
\item \code{repo_exists}: logical scalar, whether the repository exists and it is
non-empty on GitHub.
\item \code{needs_update}: whether the repository needs an update. This is
calculated from the other fields.
}
}
\description{
Also report extra refs, and whether the repository exists.
}
\details{
\code{async_get_missing_refs()} is the aynchronous version of
\code{get_missing_refs()}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_missing_refs("limma")
\dontshow{\}) # examplesIf}
}
\seealso{
Other refs: 
\code{\link{get_all_bioc_refs}()},
\code{\link{get_all_github_refs}()},
\code{\link{get_bioc_refs}()},
\code{\link{get_github_refs}()},
\code{\link{read_github_state}()},
\code{\link{write_github_state}()}
}
\concept{refs}
