% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_bioc_repos}
\alias{list_bioc_repos}
\title{List all packages in the Bioconductor git repository}
\usage{
list_bioc_repos()
}
\value{
Sorted character vector of package (=repository) names.
}
\description{
List all packages in the Bioconductor git repository
}
\details{
It works by web scraping https://git.bioconductor.org, so it'll probably
stop working if Bioconductor changes their git infrastructure.

It skips packages that are known to e problematic, typically because
of files bigger than 100MB. These files are not allowed on GitHub
}
