% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{read_github_state}
\alias{read_github_state}
\title{Read the state of the Bioconductor mirror at Github, from files}
\usage{
read_github_state(path = ".")
}
\arguments{
\item{path}{Path to the root directory of the state files.}
}
\value{
Named list of data frames, in the format of \code{\link[=get_github_refs]{get_github_refs()}}.
}
\description{
Read the state of the Bioconductor mirror at Github, from files
}
\details{
For each package, the state is in a CSV file with headers, and without
row names, at

\if{html}{\out{<div class="sourceCode">}}\preformatted{github-refs/<prefix>/<package-name>
}\if{html}{\out{</div>}}

where \verb{<refix>} is the two character, lowercase prefix of the package
name. E.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{github-refs/a4/a4
github-refs/ve/Vega
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_gituhb_state("../biocatgh-data")
\dontshow{\}) # examplesIf}
}
\seealso{
Other refs: 
\code{\link{get_all_bioc_refs}()},
\code{\link{get_all_github_refs}()},
\code{\link{get_bioc_refs}()},
\code{\link{get_github_refs}()},
\code{\link{get_missing_refs}()},
\code{\link{write_github_state}()}
}
\concept{refs}
