% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-package.R
\name{update_package}
\alias{update_package}
\title{Update one package at the Bioconductor GitHub mirror}
\usage{
update_package(pkg, state = NULL, force = FALSE, sleep = 1)
}
\arguments{
\item{pkg}{Package name.}

\item{state}{\code{NULL}, or the return value of \code{\link[=get_missing_refs]{get_missing_refs()}} for
the package. If it is \code{NULL}, and it is needed (i.e. \code{force} is not
\code{TRUE}), then it is calculated.}

\item{force}{\code{TRUE} or \code{FALSE}, whether to always (try to) update the
package withour checking the git refs in both repositories first.
Set this to \code{TRUE} if you know that an update is needed, to avoid
extra git queries.}

\item{sleep}{Number of seconds to wait after GitHub git operations,
to please GitHub secondary rate limits.}
}
\value{
Character scalar, possible values:
\itemize{
\item \code{"current"}: no updates are needed.
\item \code{"large-files"}: the repository contains large files, no update was
attempted.
\item \code{"updated"}: the repository was updated.
}
}
\description{
Update one package at the Bioconductor GitHub mirror
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
update_package("limma")
\dontshow{\}) # examplesIf}
}
