% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{write_github_state}
\alias{write_github_state}
\title{Write thee state of the Bioconductor mirror at Github into files}
\usage{
write_github_state(state, path = ".")
}
\arguments{
\item{state}{State, as returned by \code{\link[=get_all_github_refs]{get_all_github_refs()}}, a named
list of data frames.}

\item{path}{Path to the root directory of the state files.}
}
\description{
Write thee state of the Bioconductor mirror at Github into files
}
\details{
See \code{\link[=read_github_data]{read_github_data()}} for the format.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
state <- get_all_github_refs()
write_github_state(state, "../biocatgh-data")
\dontshow{\}) # examplesIf}
}
\seealso{
Other refs: 
\code{\link{get_all_bioc_refs}()},
\code{\link{get_all_github_refs}()},
\code{\link{get_bioc_refs}()},
\code{\link{get_github_refs}()},
\code{\link{get_missing_refs}()},
\code{\link{read_github_state}()}
}
\concept{refs}
