% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-all.R
\name{update_all}
\alias{update_all}
\title{Update the Bioconductor mirror at GitHub}
\usage{
update_all(sleep = 5, gh_state = NULL, bioc_state = NULL)
}
\arguments{
\item{sleep}{Number of seconds to wait after each update, to please
GitHub secondary rate limits, and be gentle on the Bioconductor git
server.}
}
\description{
The \code{BIOCATGH_GITHUB_TOKEN} environment variable must be set to the
GitHub personal access token to use to update this repository.
}
